+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/mangareader/MangaReaderGenerator.kt
@@ -1,6 +1,7 @@
 package eu.kanade.tachiyomi.multisrc.mangareader
 
 import generator.ThemeSourceData.MultiLang
+import generator.ThemeSourceData.SingleLang
 import generator.ThemeSourceGenerator
 
 class MangaReaderGenerator : ThemeSourceGenerator {
@@ -21,7 +22,16 @@ class MangaReaderGenerator : ThemeSourceGenerator {
             baseUrl = "https://mangafire.to",
             langs = listOf("en", "es", "es-419", "fr", "ja", "pt", "pt-BR"),
             isNsfw = true,
-            overrideVersionCode = 2,
+            overrideVersionCode = 3,
+        ),
+        SingleLang(
+            name = "Manhuagold",
+            baseUrl = "https://manhuagold.com",
+            lang = "en",
+            isNsfw = true,
+            className = "Manhuagold",
+            pkgName = "comickiba",
+            overrideVersionCode = 33,
         ),
     )